koji-hub: configure builder account
Before we start the kojid service on a builder, we must run the
equivalent of "koji add-host" for this builder.

(This is an bad trap for new users, because they have to delete and
re-add the account. We should update Koji to make this safer. Options:

A) Update kojid to ask "am I a host"? before trying to authenticate to a
   hub.

B) Update koji-hub and kojid to negotiate an authentication scope during
   sslLogin, so that the hub wll raise an AuthError if the account is
   not a host account.

C) Better UX, but possibly less secure: Update koji-hub to auto-accept
   any builders that match HostPrincipalFormat