koji-builder: set krb_principal explicitly
/usr/sbin/kojid calls socket.getfqdn() to discover the FQDN. This value
can be incorrect. Sometimes this will return the value of the DNS PTR
record. In cloud environments or other virtual environments, it's quite
likely that DNS records will not correctly resolve for builder hosts' IP
addresses.

Random thoughts:

Q: socket.getfqdn() is a strange choice over socket.gethostname(). Is there
   any reason why we do not use socket.gethostname() instead?
A: The Python "socket" docs say that there is no guarantee that
   "socket.gethostname()" will return a fully-qualified hostname. I
   wonder if that is still true.

Q: What happens when getfqdn() cannot get a PTR record back at all?
A: It falls back to gethostname()'s behavior and simply returns the
   uname() syscall results.

Q: Is there anything in kojid's git history about the use of getfqdn()
   or gethostname() with host_principal_format?
A: No, it was like this from the very first code drop in 2007.