kdc: stop using KEYRING credential caches
KEYRING caches leak between the container and the host system. "keyctl
show" displays the same values in on the host and inside individual
podman containers (even when I stop and start new containers). New
molecule runs will re-use the old containers' credential caches, leading
to Kerberos errors.

https://bugzilla.redhat.com/show_bug.cgi?id=1752005 explains more about
this problem in the context of FreeIPA. The IPA team's resolution was to
remove the default_ccache_name setting so that we use upstream MIT
Kerberos' default cache type (FILE) instead.

Tangentially, modern Fedoras ship a "sssd-kcm" package that includes a
/etc/krb5.conf.d/kcm_default_ccache snippet, so that will override
whatever is set in /etc/krb5.conf. However, for the current CentOS
8 environment that we use in Molecule, that package is not present, and
sssd's kcm is not necessary.